LayoutTests/inspector/sampling-profiler/named-function-expression-expected.txt


== Running test suite: ScriptProfiler.Samples.NamedFunctionExpression
-- Running test case: Sampling Profiler basic
PASS: Should have seen stacktrace:
[
  {
    "name": "bar"
  },
  {
    "name": "foo"
  },
  {
    "name": "runFor"
  }
]
